Partial refactor
