+++ components/dropdown/DropdownCaret.js
@@ -3,12 +3,12 @@ import PropTypes from 'prop-types';
 import { classnames } from '../../helpers/component';
 import Icon from '../icon/Icon';
 
-const SuperDropdownCaret = ({ isOpen }) => {
+const DropdownCaret = ({ isOpen }) => {
     return <Icon className={classnames(['expand-caret', isOpen && 'rotateX-180'])} size={12} name="caret" />;
 };
 
-SuperDropdownCaret.propTypes = {
+DropdownCaret.propTypes = {
     isOpen: PropTypes.bool
 };
 
-export default SuperDropdownCaret;
+export default DropdownCaret;
