+++ components/dropdown/ExampleDropdown.js
@@ -1,19 +1,19 @@
 import React from 'react';
-import SuperDropdown from './SuperDropdown';
+import Dropdown from './Dropdown';
 import { usePopperAnchor } from '../Popper';
-import SuperDropdownButton from './SuperDropdownButton';
+import DropdownAnchorButton from './DropdownAnchorButton';
 
 const ExampleDropdown = () => {
     const { anchorRef, isOpen, toggle, close } = usePopperAnchor();
 
     return (
         <>
-            <SuperDropdownButton buttonRef={anchorRef} isOpen={isOpen} onClick={toggle} hasCaret>
+            <DropdownAnchorButton buttonRef={anchorRef} isOpen={isOpen} onClick={toggle} hasCaret>
                 drop da bass
-            </SuperDropdownButton>
-            <SuperDropdown isOpen={isOpen} anchorRef={anchorRef} close={close}>
+            </DropdownAnchorButton>
+            <Dropdown isOpen={isOpen} anchorRef={anchorRef} close={close}>
                 <div style={{ height: 400 }}>asdfkljasdf</div>
-            </SuperDropdown>
+            </Dropdown>
         </>
     );
 };
