--- components/dropdown/index.js
@@ -1,3 +1,0 @@
-export { default as Dropdown } from './Dropdown';
-export { default as DropdownMenu } from './DropdownMenu';
-export { default as DropdownButton } from './DropdownButton';
